fancode_sports_app/requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.5
mysql-connector==2.2.9
mysql-connector-python==8.0.30
PyMySQL==1.1.0
SQLAlchemy==2.0.19
python-dotenv==1.0.0